@@ -2,11 +2,10 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, ClassVar, Protocol, TypeVar, Union
+from typing import TYPE_CHECKING, Any, ClassVar, Protocol, TypeAlias, TypeVar, Union
 
 from langgraph.graph._node import StateNode
 from pydantic import BaseModel
-from typing_extensions import TypeAlias
 
 if TYPE_CHECKING:
     from dataclasses import Field
